+++ ivy_tests/test_ivy/test_frontends/test_jax/test_jax_numpy_namespace_functions.py
@@ -11,41 +11,6 @@ from ivy_tests.test_ivy.test_functional.test_core.test_statistical import (
     statistical_dtype_values,
     _get_castable_dtype,
 )
-from ivy_tests.test_ivy.test_functional.test_core.test_linalg import (
-    _get_dtype_value1_value2_axis_for_tensordot,
-)
-
-
-# absolute
-@handle_frontend_test(
-    fn_tree="jax.numpy.absolute",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("signed_integer"),
-    ),
-)
-def test_jax_numpy_absolute(
-    *,
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        all_aliases=["numpy.abs"],
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
 
 
 # argmax
@@ -156,38 +121,6 @@ def test_jax_numpy_argwhere(
     )
 
 
-# add
-@handle_frontend_test(
-    fn_tree="jax.numpy.add",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"), num_arrays=2, shared_dtype=True
-    ),
-)
-def test_jax_numpy_add(
-    *,
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x1=x[0],
-        x2=x[0],
-    )
-
-
 # all
 @handle_frontend_test(
     fn_tree="jax.numpy.all",
@@ -217,79 +150,6 @@ def test_jax_numpy_all(
         fn_tree=fn_tree,
         on_device=on_device,
         a=x[0],
-    )
-
-
-# tan
-@handle_frontend_test(
-    fn_tree="jax.numpy.tan",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_tan(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# allclose
-@handle_frontend_test(
-    fn_tree="jax.numpy.allclose",
-    dtype_and_input=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
-        num_arrays=2,
-        shared_dtype=True,
-    ),
-    rtol=st.floats(min_value=1e-5, max_value=1e-1),
-    atol=st.floats(min_value=1e-8, max_value=1e-6),
-    equal_nan=st.booleans(),
-)
-def test_jax_numpy_allclose(
-    *,
-    dtype_and_input,
-    equal_nan,
-    rtol,
-    atol,
-    as_variable,
-    with_out,
-    num_positional_args,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, input = dtype_and_input
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=with_out,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        rtol=rtol,
-        atol=atol,
-        a=input[0],
-        b=input[1],
-        equal_nan=equal_nan,
     )
 
 
@@ -556,103 +416,6 @@ def test_jax_numpy_var(
     )
 
 
-# dot
-@st.composite
-def _get_dtype_input_and_vectors(draw):
-    dim_size = draw(helpers.ints(min_value=1, max_value=5))
-    dtype = draw(helpers.get_dtypes("float", index=1, full=False))
-    if dim_size == 1:
-        vec1 = draw(
-            helpers.array_values(
-                dtype=dtype[0], shape=(dim_size,), min_value=2, max_value=5
-            )
-        )
-        vec2 = draw(
-            helpers.array_values(
-                dtype=dtype[0], shape=(dim_size,), min_value=2, max_value=5
-            )
-        )
-    else:
-        vec1 = draw(
-            helpers.array_values(
-                dtype=dtype[0], shape=(dim_size, dim_size), min_value=2, max_value=5
-            )
-        )
-        vec2 = draw(
-            helpers.array_values(
-                dtype=dtype[0], shape=(dim_size, dim_size), min_value=2, max_value=5
-            )
-        )
-    return dtype, vec1, vec2
-
-
-# mod
-@handle_frontend_test(
-    fn_tree="jax.numpy.mod",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        num_arrays=2,
-        shared_dtype=True,
-    ),
-)
-def test_jax_numpy_mod(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    assume(not np.any(np.isclose(x[1], 0)))
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
-@handle_frontend_test(
-    fn_tree="jax.numpy.dot",
-    dtype_x_y=_get_dtype_input_and_vectors(),
-)
-def test_jax_numpy_dot(
-    *,
-    dtype_x_y,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x, y = dtype_x_y
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        rtol=1e-01,
-        atol=1e-01,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a=x,
-        b=y,
-        precision=None,
-    )
-
-
 # einsum
 @handle_frontend_test(
     fn_tree="jax.numpy.einsum",
@@ -705,244 +468,6 @@ def test_jax_numpy_einsum(
     )
 
 
-# arctan
-@handle_frontend_test(
-    fn_tree="jax.numpy.arctan",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_arctan(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    with_out,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=with_out,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# arctan2
-@handle_frontend_test(
-    fn_tree="jax.numpy.arctan2",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
-        num_arrays=2,
-        shared_dtype=True,
-    ),
-)
-def test_jax_numpy_arctan2(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    with_out,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=with_out,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
-@handle_frontend_test(
-    fn_tree="jax.numpy.cos",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_cos(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    with_out,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=with_out,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# cosh
-@handle_frontend_test(
-    fn_tree="jax.numpy.cosh",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_cosh(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# tanh
-@handle_frontend_test(
-    fn_tree="jax.numpy.tanh",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_tanh(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# sinh
-@handle_frontend_test(
-    fn_tree="jax.numpy.sinh",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_sinh(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-@handle_frontend_test(
-    fn_tree="jax.numpy.sin",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_sin(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# floor
-@handle_frontend_test(
-    fn_tree="jax.numpy.floor",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_floor(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
 # fmax
 @handle_frontend_test(
     fn_tree="jax.numpy.fmax",
@@ -976,77 +501,6 @@ def test_jax_numpy_fmax(
         on_device=on_device,
         x1=inputs[0],
         x2=inputs[1],
-    )
-
-
-@handle_frontend_test(
-    fn_tree="jax.numpy.array_equal",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        num_arrays=2,
-        min_value=-np.inf,
-        max_value=np.inf,
-        shared_dtype=True,
-    ),
-    equal_nan=st.booleans(),
-)
-def test_jax_numpy_array_equal(
-    *,
-    dtype_and_x,
-    equal_nan,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a1=x[0],
-        a2=x[1],
-        equal_nan=equal_nan,
-    )
-
-
-@handle_frontend_test(
-    fn_tree="jax.numpy.array_equiv",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        num_arrays=2,
-        shared_dtype=True,
-    ),
-)
-def test_jax_numpy_array_equiv(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a1=x[0],
-        a2=x[1],
     )
 
 
@@ -1084,122 +538,6 @@ def test_jax_numpy_zeros(
         on_device=on_device,
         shape=shape,
         dtype=dtypes[0],
-    )
-
-
-# arccos
-@handle_frontend_test(
-    fn_tree="jax.numpy.arccos",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_arccos(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# arccosh
-@handle_frontend_test(
-    fn_tree="jax.numpy.arccosh",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_arccosh(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# arcsin
-@handle_frontend_test(
-    fn_tree="jax.numpy.arcsin",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_arcsin(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# arcsinh
-@handle_frontend_test(
-    fn_tree="jax.numpy.arcsinh",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_arcsinh(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
     )
 
 
@@ -1371,38 +709,6 @@ def test_jax_numpy_bitwise_xor(
     )
 
 
-# power
-@handle_frontend_test(
-    fn_tree="jax.numpy.power",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"), num_arrays=2, shared_dtype=True
-    ),
-)
-def test_jax_numpy_power(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
 # arange
 @handle_frontend_test(
     fn_tree="jax.numpy.arange",
@@ -1528,103 +834,6 @@ def test_jax_numpy_cumprod(
     )
 
 
-# trunc
-@handle_frontend_test(
-    fn_tree="jax.numpy.trunc",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_trunc(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-@handle_frontend_test(
-    fn_tree="jax.numpy.ceil",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_ceil(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# float_power
-@handle_frontend_test(
-    fn_tree="jax.numpy.float_power",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
-        min_value=-10,
-        max_value=10,
-        num_arrays=2,
-        shared_dtype=True,
-        min_num_dims=1,
-        max_num_dims=3,
-        min_dim_size=1,
-        max_dim_size=3,
-    ),
-)
-def test_jax_numpy_float_power(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
 # cumsum
 @handle_frontend_test(
     fn_tree="jax.numpy.cumsum",
@@ -1708,269 +917,6 @@ def test_jax_numpy_heaviside(
     )
 
 
-# deg2rad
-@handle_frontend_test(
-    fn_tree="jax.numpy.deg2rad",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
-    ),
-)
-def test_jax_numpy_deg2rad(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# exp2
-@handle_frontend_test(
-    fn_tree="jax.numpy.exp2",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
-        min_value=-10,
-        max_value=10,
-        min_num_dims=1,
-        max_num_dims=3,
-        min_dim_size=1,
-        max_dim_size=3,
-    ),
-)
-def test_jax_numpy_exp2(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-        rtol=1e-01,
-        atol=1e-02,
-    )
-
-
-# gcd
-@handle_frontend_test(
-    fn_tree="jax.numpy.gcd",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("integer"),
-        min_value=-100,
-        max_value=100,
-        min_num_dims=1,
-        max_num_dims=3,
-        min_dim_size=1,
-        max_dim_size=3,
-        num_arrays=2,
-        shared_dtype=True,
-    ),
-)
-def test_jax_numpy_gcd(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
-# i0
-@handle_frontend_test(
-    fn_tree="jax.numpy.i0",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
-        min_value=-10,
-        max_value=10,
-        min_num_dims=1,
-        max_num_dims=3,
-        min_dim_size=1,
-        max_dim_size=3,
-    ),
-)
-def test_jax_numpy_i0(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# isneginf
-@handle_frontend_test(
-    fn_tree="jax.numpy.isneginf",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
-        min_value=-np.inf,
-        max_value=np.inf,
-        min_num_dims=1,
-        max_num_dims=3,
-        min_dim_size=1,
-        max_dim_size=3,
-        allow_inf=True,
-    ),
-)
-def test_jax_numpy_isneginf(
-    *,
-    dtype_and_x,
-    with_out,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=with_out,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# isposinf
-@handle_frontend_test(
-    fn_tree="jax.numpy.isposinf",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
-        min_value=-np.inf,
-        max_value=np.inf,
-        min_num_dims=1,
-        max_num_dims=3,
-        min_dim_size=1,
-        max_dim_size=3,
-        allow_inf=True,
-    ),
-)
-def test_jax_numpy_isposinf(
-    *,
-    dtype_and_x,
-    with_out,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=with_out,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# kron
-@handle_frontend_test(
-    fn_tree="jax.numpy.kron",
-    dtype_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        min_num_dims=1,
-        max_num_dims=3,
-        min_dim_size=1,
-        max_dim_size=3,
-        num_arrays=2,
-        shared_dtype=True,
-    ),
-)
-def test_jax_numpy_kron(
-    *,
-    dtype_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a=x[0],
-        b=x[1],
-    )
-
-
 # sum
 @handle_frontend_test(
     fn_tree="jax.numpy.sum",
@@ -2022,167 +968,6 @@ def test_jax_numpy_sum(
         keepdims=keepdims,
         initial=initial,
         where=where,
-    )
-
-
-# lcm
-@handle_frontend_test(
-    fn_tree="jax.numpy.lcm",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("integer"),
-        num_arrays=2,
-        shared_dtype=True,
-        min_num_dims=1,
-        max_num_dims=3,
-        min_value=-100,
-        max_value=100,
-        allow_nan=False,
-    ),
-)
-def test_jax_numpy_lcm(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
-# logaddexp2
-@handle_frontend_test(
-    fn_tree="jax.numpy.logaddexp2",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
-        num_arrays=2,
-        shared_dtype=True,
-        min_num_dims=1,
-        max_num_dims=3,
-        min_value=-100,
-        max_value=100,
-        allow_nan=False,
-    ),
-)
-def test_jax_numpy_logaddexp2(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        rtol=1e-01,
-        atol=1e-02,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
-# trapz
-@st.composite
-def _either_x_dx(draw):
-    dtype_values_axis = draw(
-        helpers.dtype_values_axis(
-            available_dtypes=st.shared(helpers.get_dtypes("float"), key="trapz_dtype"),
-            min_value=-100,
-            max_value=100,
-            min_num_dims=1,
-            max_num_dims=3,
-            min_dim_size=1,
-            max_dim_size=3,
-            allow_neg_axes=True,
-            valid_axis=True,
-            force_int_axis=True,
-        ),
-    )
-    rand = (draw(st.integers(min_value=0, max_value=1)),)
-    if rand == 0:
-        either_x_dx = draw(
-            helpers.dtype_and_x(
-                avaliable_dtypes=st.shared(
-                    helpers.get_dtypes("float"), key="trapz_dtype"
-                ),
-                min_value=-100,
-                max_value=100,
-                min_num_dims=1,
-                max_num_dims=3,
-                min_dim_size=1,
-                max_dim_size=3,
-            )
-        )
-        return dtype_values_axis, rand, either_x_dx
-    else:
-        either_x_dx = draw(
-            st.floats(min_value=-10, max_value=10),
-        )
-        return dtype_values_axis, rand, either_x_dx
-
-
-@handle_frontend_test(
-    fn_tree="jax.numpy.trapz",
-    dtype_x_axis_rand_either=_either_x_dx(),
-)
-def test_jax_numpy_trapz(
-    *,
-    dtype_x_axis_rand_either,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    dtype_values_axis, rand, either_x_dx = dtype_x_axis_rand_either
-    input_dtype, y, axis = dtype_values_axis
-    if rand == 0:
-        dtype_x, x = either_x_dx
-        x = np.asarray(x, dtype=dtype_x)
-        dx = None
-    else:
-        x = None
-        dx = either_x_dx
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        rtol=1e-2,
-        atol=1e-2,
-        y=y[0],
-        x=x,
-        dx=dx,
-        axis=axis,
     )
 
 
@@ -2289,35 +1074,6 @@ def test_jax_numpy_diag(
     )
 
 
-# sqrt
-@handle_frontend_test(
-    fn_tree="jax.numpy.sqrt",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_sqrt(
-    *,
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
 @handle_frontend_test(
     fn_tree="jax.numpy.hstack",
     dtype_and_tup=helpers.dtype_and_values(
@@ -2347,35 +1103,6 @@ def test_jax_numpy_hstack(
         frontend=frontend,
         fn_tree=fn_tree,
         tup=x,
-    )
-
-
-# arctanh
-@handle_frontend_test(
-    fn_tree="jax.numpy.arctanh",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
-        min_num_dims=0,
-    ),
-)
-def test_jax_numpy_arctanh(
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        x=x[0],
     )
 
 
@@ -2471,223 +1198,6 @@ def test_jax_numpy_msort(
         frontend=frontend,
         fn_tree=fn_tree,
         a=x[0],
-    )
-
-
-# multiply
-@handle_frontend_test(
-    fn_tree="jax.numpy.multiply",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        num_arrays=2,
-        shared_dtype=True,
-    ),
-)
-def test_jax_numpy_multiply(
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
-# not_equal
-@handle_frontend_test(
-    fn_tree="jax.numpy.not_equal",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        num_arrays=2,
-        shared_dtype=True,
-    ),
-)
-def test_jax_numpy_not_equal(
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
-# less
-@handle_frontend_test(
-    fn_tree="jax.numpy.less",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        num_arrays=2,
-        shared_dtype=True,
-    ),
-)
-def test_jax_numpy_less(
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
-# less_equal
-@handle_frontend_test(
-    fn_tree="jax.numpy.less_equal",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        num_arrays=2,
-        shared_dtype=True,
-    ),
-)
-def test_jax_numpy_less_equal(
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
-# greater
-@handle_frontend_test(
-    fn_tree="jax.numpy.greater",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        num_arrays=2,
-        shared_dtype=True,
-    ),
-)
-def test_jax_numpy_greater(
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
-# greater_equal
-@handle_frontend_test(
-    fn_tree="jax.numpy.greater_equal",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        num_arrays=2,
-        shared_dtype=True,
-    ),
-)
-def test_jax_numpy_greater_equal(
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
-# equal
-@handle_frontend_test(
-    fn_tree="jax.numpy.equal",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        num_arrays=2,
-        shared_dtype=True,
-    ),
-)
-def test_jax_numpy_equal(
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        x1=x[0],
-        x2=x[1],
     )
 
 
@@ -2787,84 +1297,6 @@ def test_jax_numpy_max(
     )
 
 
-# log10
-@handle_frontend_test(
-    fn_tree="jax.numpy.log10",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
-        min_num_dims=1,
-        max_num_dims=3,
-        min_value=-100,
-        max_value=100,
-    ),
-)
-def test_jax_numpy_log10(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        rtol=1e-01,
-        atol=1e-02,
-        x=x[0],
-    )
-
-
-# logaddexp
-@handle_frontend_test(
-    fn_tree="jax.numpy.logaddexp",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
-        num_arrays=2,
-        shared_dtype=True,
-        min_num_dims=1,
-        max_num_dims=3,
-        min_value=-100,
-        max_value=100,
-        allow_nan=False,
-    ),
-)
-def test_jax_numpy_logaddexp(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        rtol=1e-01,
-        atol=1e-02,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
 @st.composite
 def dims_and_offset(draw, shape):
     shape_actual = draw(shape)
@@ -2921,37 +1353,6 @@ def test_jax_numpy_diagonal(
         offset=offset,
         axis1=axis1,
         axis2=axis2,
-    )
-
-
-# degrees
-@handle_frontend_test(
-    fn_tree="jax.numpy.degrees",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("float"),
-    ),
-)
-def test_jax_numpy_degrees(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
     )
 
 
@@ -3067,158 +1468,4 @@ def test_numpy_zeros_like(
         a=x[0],
         dtype=dtype[0],
         shape=shape,
-    )
-
-
-@handle_frontend_test(
-    fn_tree="jax.numpy.negative",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"), min_num_dims=1
-    ),
-)
-def test_jax_numpy_negative(
-    dtype_and_x,
-    as_variable,
-    with_out,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-    on_device,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-@handle_frontend_test(
-    fn_tree="jax.numpy.rad2deg",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"), min_num_dims=1
-    ),
-)
-def test_jax_numpy_rad2deg(
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-    on_device,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# tensordot
-@handle_frontend_test(
-    fn_tree="jax.numpy.tensordot",
-    dtype_values_and_axes=_get_dtype_value1_value2_axis_for_tensordot(
-        helpers.get_dtypes(kind="numeric")
-    ),
-)
-def test_jax_numpy_tensordot(
-    dtype_values_and_axes,
-    as_variable,
-    native_array,
-    num_positional_args,
-    frontend,
-    fn_tree,
-):
-    dtype, a, b, axes = dtype_values_and_axes
-    helpers.test_frontend_function(
-        input_dtypes=dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        a=a,
-        b=b,
-        axes=axes,
-    )
-
-
-# divide
-@handle_frontend_test(
-    fn_tree="jax.numpy.divide",
-    dtype_values=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        num_arrays=2,
-        allow_inf=False,
-        large_abs_safety_factor=4,
-        safety_factor_scale="linear",
-        shared_dtype=True,
-    ),
-)
-def test_jax_numpy_divide(
-    *,
-    dtype_values,
-    as_variable,
-    native_array,
-    num_positional_args,
-    frontend,
-    fn_tree,
-):
-    input_dtype, x = dtype_values
-    assume(not np.any(np.isclose(x[1], 0)))
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        all_aliases=["numpy.true_divide"],
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        a=x[0],
-        b=x[1],
-    )
-
-
-# exp
-@handle_frontend_test(
-    fn_tree="jax.numpy.exp",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-)
-def test_jax_numpy_exp(
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
     )
